+++ Nos/Views/Home/HomeFeedView.swift
@@ -11,9 +11,7 @@ struct HomeFeedView: View {
     @Environment(CurrentUser.self) var currentUser
     @ObservationIgnored @Dependency(\.analytics) private var analytics
 
-    @State private var lastRefreshDate = Date(
-        timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.staticLoadTime)
-    )
+    @State private var refreshController: RefreshController = DefaultRefreshController()
     @State private var isVisible = false
     @State private var relaySubscriptions = [SubscriptionCancellable]()
     @State private var isShowingRelayList = false
@@ -36,9 +34,22 @@ struct HomeFeedView: View {
     }
     
     var homeFeedFetchRequest: NSFetchRequest<Event> {
-        Event.homeFeed(for: user, before: lastRefreshDate, seenOn: selectedRelay) 
+        Event.homeFeed(
+            for: user,
+            before: refreshController.lastRefreshDate ??
+                Date(timeIntervalSince1970: Date.now.timeIntervalSince1970 + Double(Self.staticLoadTime)),
+            seenOn: selectedRelay
+        )
     }
-    
+
+    var newNotesRequest: NSFetchRequest<Event> {
+        Event.homeFeed(
+            for: user,
+            after: refreshController.lastRefreshDate ?? .now,
+            seenOn: selectedRelay
+        )
+    }
+
     var homeFeedFilter: Filter {
         var filter = Filter(kinds: [.text, .delete, .repost, .longFormContent, .report])
         if selectedRelay == nil {
@@ -61,12 +72,13 @@ struct HomeFeedView: View {
                 databaseFilter: homeFeedFetchRequest,
                 relayFilter: homeFeedFilter,
                 relay: selectedRelay,
-                context: viewContext,
+                managedObjectContext: viewContext,
                 tab: .home,
+                refreshController: refreshController,
                 header: {
                     EmptyView()
                 },
-                emptyPlaceholder: { _ in
+                emptyPlaceholder: {
                     VStack {
                         Text(.localizable.noEvents)
                             .padding()
@@ -74,12 +86,15 @@ struct HomeFeedView: View {
                     .frame(minHeight: 300)
                 },
                 onRefresh: {
-                    lastRefreshDate = .now
-                    return Event.homeFeed(for: user, before: lastRefreshDate)
+                    refreshController.setLastRefreshDate(.now)
                 }
             )
             .padding(0)
-            
+
+            NewNotesButton(fetchRequest: FetchRequest(fetchRequest: newNotesRequest)) {
+                refreshController.setShouldRefresh(true)
+            }
+
             if showTimedLoadingIndicator {
                 FullscreenProgressView(
                     isPresented: $showTimedLoadingIndicator,
@@ -96,7 +111,7 @@ struct HomeFeedView: View {
                 )
                 .onChange(of: selectedRelay) { _, _ in
                     showTimedLoadingIndicator = true
-                    lastRefreshDate = .now + Self.staticLoadTime
+                    refreshController.setLastRefreshDate(.now + Self.staticLoadTime)
                     Task {
                         withAnimation {
                             showRelayPicker = false